bot/requirements.txt
pymongo==3.4.0
beautifulsoup4==4.5.1
python-telegram-bot==5.3.0
requests==2.13.0
